+++ packages/abstract-provider/lib.esm/index.d.ts
@@ -73,6 +73,7 @@ export interface TransactionReceipt {
     blockNumber: number;
     confirmations: number;
     cumulativeGasUsed: BigNumber;
+    effectiveGasPrice: BigNumber;
     byzantium: boolean;
     type: number;
     status?: number;