+++ packages/providers/lib.esm/etherscan-provider.js
@@ -25,8 +25,11 @@ function getTransactionPostData(transaction) {
             continue;
         }
         let value = transaction[key];
+        if (key === "type" && value === 0) {
+            continue;
+        }
         // Quantity-types require no leading zero, unless 0
-        if ({ type: true, gasLimit: true, gasPrice: true, nonce: true, value: true }[key]) {
+        if ({ type: true, gasLimit: true, gasPrice: true, maxFeePerGs: true, maxPriorityFeePerGas: true, nonce: true, value: true }[key]) {
             value = hexValue(hexlify(value));
         }
         else if (key === "accessList") {