+++ packages/providers/lib.esm/formatter.js
@@ -87,6 +87,7 @@ export class Formatter {
             blockNumber: number,
             confirmations: Formatter.allowNull(number, null),
             cumulativeGasUsed: bigNumber,
+            effectiveGasPrice: Formatter.allowNull(bigNumber),
             status: Formatter.allowNull(number),
             type: type
         };
@@ -276,11 +277,6 @@ export class Formatter {
             transaction.accessList = [];
         }
         const result = Formatter.check(this.formats.transaction, transaction);
-        if (result.type === 2) {
-            if (result.gasPrice == null) {
-                result.gasPrice = result.maxFeePerGas;
-            }
-        }
         if (transaction.chainId != null) {
             let chainId = transaction.chainId;
             if (isHexString(chainId)) {