+++ packages/testcases/lib/testcases.d.ts
@@ -77,6 +77,28 @@ export interface SignedTransaction {
     value: string;
     data: string;
 }
+export interface TypedTransaction {
+    name: string;
+    key: string;
+    address: string;
+    tx: {
+        type?: number;
+        data?: string;
+        gasLimit?: string;
+        maxPriorityFeePerGas: string;
+        maxFeePerGas: string;
+        nonce: number;
+        to: string;
+        value: string;
+        chainId: number;
+        accessList: Array<{
+            address: string;
+            storageKeys: Array<string>;
+        }>;
+    };
+    signed: string;
+    unsigned: string;
+}
 export interface Eip712 {
     name: string;
     domain: {