+++ packages/tests/lib.esm/test-utils.js
@@ -501,6 +501,94 @@ describe("Test Signature Manipulation", function () {
         });
     });
 });
+describe("Test Typed Transactions", function () {
+    const tests = loadTests("typed-transactions");
+    function equalsData(name, a, b, ifNull) {
+        assert.equal(ethers.utils.hexlify(a), ethers.utils.hexlify((b == null) ? ifNull : b), name);
+        return true;
+    }
+    function equalsNumber(name, a, b, ifNull) {
+        assert.ok(ethers.BigNumber.from(a).eq((b == null) ? ifNull : b), name);
+        return true;
+    }
+    function equalsArray(name, a, b, equals) {
+        assert.equal(a.length, b.length, `${name}.length`);
+        for (let i = 0; i < a.length; i++) {
+            if (!equals(`${name}[${i}]`, a[i], b[i])) {
+                return false;
+            }
+        }
+        return true;
+    }
+    function makeEqualsArray(equals) {
+        return function (name, a, b) {
+            return equalsArray(name, a, b, equals);
+        };
+    }
+    function equalsAccessList(name, a, b) {
+        return equalsArray(`${name}-address`, a.map((f) => f.address), b.map((f) => f.address), equalsData) &&
+            equalsArray(`${name}-storageKeys`, a.map((f) => f.storageKeys), b.map((f) => f.storageKeys), makeEqualsArray(equalsData));
+    }
+    function allowNull(name, a, b, equals) {
+        if (a == null) {
+            assert.ok(b == null, `${name}:!NULL`);
+            return true;
+        }
+        else if (b == null) {
+            assert.fail(`${name}:!!NULL`);
+        }
+        return equals(name, a, b);
+    }
+    function equalsCommonTransaction(name, a, b) {
+        return equalsNumber(`${name}-type`, a.type, b.type, 0) &&
+            equalsData(`${name}-data`, a.data, b.data, "0x") &&
+            equalsNumber(`${name}-gasLimit`, a.gasLimit, b.gasLimit, 0) &&
+            equalsNumber(`${name}-nonce`, a.nonce, b.nonce, 0) &&
+            allowNull(`${name}-to`, a.to, b.to, equalsData) &&
+            equalsNumber(`${name}-value`, a.value, b.value, 0) &&
+            equalsNumber(`${name}-chainId`, a.chainId, b.chainId, 0) &&
+            equalsAccessList(`${name}-accessList`, a.accessList, b.accessList || []);
+    }
+    function equalsEip1559Transaction(name, a, b) {
+        return equalsNumber(`${name}-maxPriorityFeePerGas`, a.maxPriorityFeePerGas, b.maxPriorityFeePerGas, 0) &&
+            equalsNumber(`${name}-maxFeePerGas`, a.maxFeePerGas, b.maxFeePerGas, 0) &&
+            equalsCommonTransaction(name, a, b);
+    }
+    function equalsEip2930Transaction(name, a, b) {
+        return equalsNumber(`${name}-gasPrice`, a.gasPrice, b.gasPrice, 0) &&
+            equalsCommonTransaction(name, a, b);
+    }
+    function equalsTransaction(name, a, b) {
+        switch (a.type) {
+            case 1:
+                return equalsEip2930Transaction(name, a, b);
+            case 2:
+                return equalsEip1559Transaction(name, a, b);
+        }
+        assert.fail(`unknown transaction type ${a.type}`);
+    }
+    tests.forEach((test, index) => {
+        it(test.name, function () {
+            return __awaiter(this, void 0, void 0, function* () {
+                {
+                    const wallet = new ethers.Wallet(test.key);
+                    const signed = yield wallet.signTransaction(test.tx);
+                    assert.equal(signed, test.signed, "signed transactions match");
+                }
+                assert.equal(ethers.utils.serializeTransaction(test.tx), test.unsigned, "unsigned transactions match");
+                {
+                    const tx = ethers.utils.parseTransaction(test.unsigned);
+                    assert.ok(equalsTransaction("transaction", tx, test.tx), "all unsigned keys match");
+                }
+                {
+                    const tx = ethers.utils.parseTransaction(test.signed);
+                    assert.ok(equalsTransaction("transaction", tx, test.tx), "all signed keys match");
+                    assert.equal(tx.from.toLowerCase(), test.address, "sender matches");
+                }
+            });
+        });
+    });
+});
 describe("BigNumber", function () {
     const tests = loadTests("bignumber");
     tests.forEach((test) => {